#49 Environment randomization
Moved mirroring logic to TREnvironmentEditor.
