+++ TREnvironmentEditor/TREnvironmentEditor.csproj
@@ -8,6 +8,7 @@
   <ItemGroup>
     <ProjectReference Include="..\TRFDControl\TRFDControl.csproj" />
     <ProjectReference Include="..\TRLevelReader\TRLevelReader.csproj" />
+    <ProjectReference Include="..\TRModelTransporter\TRModelTransporter.csproj" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
